add jwt extended features

+++ resources/item.py
@@ -1,5 +1,11 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required, 
+    get_jwt_claims, 
+    jwt_optional, 
+    get_jwt_identity,
+    fresh_jwt_required
+)
 from models.item import ItemModel
 
 class Item(Resource):
@@ -26,7 +32,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'},404
 
-    @jwt_required()
+    @jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'error': "An item with name {} already exists".format(name)}, 400
@@ -38,7 +44,7 @@ class Item(Resource):
             return {'message': 'Error inserting item'}, 500
         return item.json(), 201
 
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
@@ -50,15 +56,26 @@ class Item(Resource):
         item.save_to_db()
         return item.json(), 201
 
-    @jwt_required()
+    @fresh_jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete()
         return {'message': 'Item deleted'}
         
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
+        user_id = get_jwt_identity()
         items = ItemModel.get_all()
-        return {'items': [i.json() for i in items ]}
-        # list(map(lambda i: i.json(), items))
+        # If logged in
+        if user_id:
+            return {'items': [i.json() for i in items ]}, 200
+            # list(map(lambda i: i.json(), items))
+        return {
+            'items': [i.name for i in items],
+            'message': 'More info if logged in'
+        }, 200
